[Release 1.1] Revert Async changes in SNIPacket to fix deadlock issues + Update SNI Package Reference (#425)
