+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft.Data.SqlClient.csproj
@@ -84,7 +84,6 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="Microsoft.Data.SqlClientKey.snk" />
-    <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Microsoft\Data\Interop\SNINativeMethodWrapper.cs" />
@@ -340,8 +339,14 @@
     </COMReference>
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="System.Security.Cryptography.Algorithms">
+      <Version>4.3.1</Version>
+    </PackageReference>
+    <PackageReference Include="System.Security.Cryptography.Primitives">
+      <Version>4.3.0</Version>
+    </PackageReference>
     <PackageReference Include="Microsoft.Data.SqlClient.SNI">
-      <Version>1.1.0</Version>
+      <Version>[1.1.0,1.2.0)</Version>
     </PackageReference>
     <PackageReference Include="Microsoft.Identity.Client">
       <Version>3.0.8</Version>
