--- src/Microsoft.Data.SqlClient/netfx/src/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.Security.Cryptography.Algorithms" version="4.3.1" targetFramework="net46" />
-  <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net46" />
-  <package id="Microsoft.Data.SqlClient.SNI" version="1.1.0" targetFramework="net46" />
-</packages>
